Re-factor monitors and monitor states
There was a fair bit of code duplication between the two monitor
applications state tracking. Both applications have now moved under
the piwheels.monitor namespace, and piwheels.monitor.states contains the
state tracking code common to both monitor applications.
